+++ webserver/scenes/management/commands/legacy_migration/__init__.py
@@ -1,0 +1,2 @@
+from .IssueLog import IssueLog
+from .translate import ItemMigrator
